cloud-functions/transcript-summarizer/requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
cryptography==42.0.5
deprecation==2.1.0
Flask==3.0.2
functions-framework==3.5.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-aiplatform==1.44.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-oauth==1.0.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.1
zipp==3.18.1
